# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CoronaMecca API",
        "description": "CoronaMecca API.",
        "contact": {
            "email": "dann4624@edu.sde.dk"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://vacapi.semeicardia.online/api"
        }
    ],
    "paths": {
...
        "/zoneLogs": {
            "get": {
                "tags": [
                    "Zone Logs"
                ],
                "summary": "Get a list of Zone Logs",
                "description": "Get a list of Zone Logs",
                "operationId": "ZoneLogsList",
                "responses": {
                    "200": {
                        "description": "List of Zone Logs"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Zone Logs"
                ],
                "summary": "Create an Zone Log",
                "description": "Create an Zone Log",
                "operationId": "ZoneLogsCreate",
                "parameters": [
                    {
                        "name": "zone_id",
                        "in": "query",
                        "description": "ID of the Zone",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "log_action_id",
                        "in": "query",
                        "description": "ID of the Action",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "temperature",
                        "in": "query",
                        "description": "Temperature",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": "14.2"
                        }
                    },
                    {
                        "name": "humidity",
                        "in": "query",
                        "description": "Humidity",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": "60"
                        }
                    },
                    {
                        "name": "alarm",
                        "in": "query",
                        "description": "Alarm Triggered",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "maximum": 1,
                            "minimum": 0,
                            "example": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Zone Log created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/zoneLogs/deleted": {
            "get": {
                "tags": [
                    "Zone Logs"
                ],
                "summary": "Get a list of deleted Zone Logs",
                "description": "Get a list of deleted Zone Logs",
                "operationId": "ZoneLogsListDeleted",
                "responses": {
                    "200": {
                        "description": "List of deleted Zone Logs"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/zoneLogs/zone/{id}": {
            "get": {
                "tags": [
                    "Zone Logs"
                ],
                "summary": "Get a specific Zone's logs",
                "description": "Get a specific Zone's logs",
                "operationId": "ZoneLogsZone",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Zone Log",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Zone Log object"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/zoneLogs/{id}": {
            "get": {
                "tags": [
                    "Zone Logs"
                ],
                "summary": "Get a specific Zone Log",
                "description": "Get a specific Zone Log",
                "operationId": "ZoneLogsShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Zone Log",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Zone Log object"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Zone Logs"
                ],
                "summary": "Update an Zone Log",
                "description": "Update an Zone Log",
                "operationId": "ZoneLogsUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Zone Log",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "zone_id",
                        "in": "query",
                        "description": "ID of the Zone",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "log_action_id",
                        "in": "query",
                        "description": "ID of the Action",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "temperature",
                        "in": "query",
                        "description": "Temperature",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": "14.2"
                        }
                    },
                    {
                        "name": "humidity",
                        "in": "query",
                        "description": "Humidity",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": "60"
                        }
                    },
                    {
                        "name": "alarm",
                        "in": "query",
                        "description": "Alarm Triggered",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "maximum": 1,
                            "minimum": 0,
                            "example": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Zone Log updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Zone Logs"
                ],
                "summary": "Delete an Zone Log",
                "description": "Delete an Zone Log",
                "operationId": "ZoneLogsDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Zone Log",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Zone Log deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/zoneLogs/{id}/force": {
            "delete": {
                "tags": [
                    "Zone Logs"
                ],
                "summary": "Permanently Delete an Zone Log",
                "description": "Permanently Delete an Zone Log",
                "operationId": "ZoneLogsDeleteForce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Zone Log",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Zone Log permanently deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/zoneLogs/{id}/restore": {
            "put": {
                "tags": [
                    "Zone Logs"
                ],
                "summary": "Restore a deleted Zone Log",
                "description": "Restore a deleted Zone Log",
                "operationId": "ZoneLogsRestore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Zone Log",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Zone Log restored"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}